fuel-tracker-backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
bleach==6.2.0
coverage==7.6.9
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.11
python-decouple==3.8
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
